+++ s4e-backend/src/main/java/pl/cyfronet/s4e/controller/response/BasicProductResponse.java
@@ -1,9 +1,14 @@
 package pl.cyfronet.s4e.controller.response;
 
+import org.springframework.beans.factory.annotation.Value;
+
 public interface BasicProductResponse {
     Long getId();
 
     String getName();
 
     String getDisplayName();
+
+    @Value("#{@productService.isFavourite(target.id)}")
+    boolean isFavourite();
 }