+++ s4e-backend/src/main/java/pl/cyfronet/s4e/data/repository/ProductRepository.java
@@ -1,5 +1,6 @@
 package pl.cyfronet.s4e.data.repository;
 
+import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.transaction.annotation.Transactional;
 import pl.cyfronet.s4e.bean.Product;
@@ -12,6 +13,11 @@ public interface ProductRepository extends CrudRepository<Product, Long> {
     Optional<Product> findByNameContainingIgnoreCase(String name);
 
     <T> List<T> findAllBy(Class<T> projection);
+
+    @Query("SELECT CASE WHEN COUNT(p)> 0 THEN 'true' ELSE 'false' END FROM Product p " +
+            "JOIN p.favourites f " +
+            "WHERE p.id = :productId AND f.email = :email")
+    boolean isFavouriteByEmailAndProductId(String email, Long productId);
 
     <T> Optional<T> findById(Long id, Class<T> projection);